+++ Sources/IldamSDK/Models/Responses/NotificationsResponse.swift
@@ -16,7 +16,7 @@ public struct NotificationsResponse: Sendable {
 }
 
 public struct NotificationItem: Sendable {
-    public let id: Int?
+    public let id: Int64?
     public let title: String?
     public let type: String?
     public let content: String?
@@ -24,7 +24,7 @@ public struct NotificationItem: Sendable {
     public let readed: Bool?
     public let image: String?
     
-    public init(id: Int, title: String, type: String, content: String, createdAt: Int, readed: Bool, image: String?) {
+    public init(id: Int64, title: String, type: String, content: String, createdAt: Int, readed: Bool, image: String?) {
         self.id = id
         self.title = title
         self.type = type
